--- src/place/naverMap/map.dto.ts
@@ -1,1 +1,0 @@
-export class MapDto {}
