--- src/place/naverSearch/search.dto.ts
@@ -1,9 +1,0 @@
-import { IsEmail, IsNotEmpty } from "class-validator";
-
-export class SearchDto {
-  @IsNotEmpty()
-  readonly query: string;
-
-  readonly display: number;
-  readonly sort: string;
-}
