+++ src/place/place.controller.ts
@@ -1,18 +1,18 @@
 import { Controller, Get, Query } from "@nestjs/common";
-import { PlaceService } from "./naverSearch/search.service";
-import { SearchDto } from "./naverSearch/search.dto";
+import { PlaceService } from "./kakaoMap/search.service";
+import { KeywordSearchDto } from "./kakaoMap/search.dto";
 
 @Controller("place")
 export class PlaceController {
   constructor(private readonly placeService: PlaceService) {}
 
   @Get("/search")
-  async getPlace(@Query() p: SearchDto): Promise<any> {
+  async getPlaceByKeyworld(
+    @Query() keywordSearchDto: KeywordSearchDto
+  ): Promise<any> {
     // TODO: mongodb cache 추가
-    const searchPlaces = await this.placeService.findAll(
-      p.query,
-      p.display,
-      p.sort
+    const searchPlaces = await this.placeService.searchByKeyworld(
+      keywordSearchDto
     );
 
     // TODO: api wrap with graphql
