+++ src/place/place.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { ConfigModule } from "../config/config.module";
 
-import { PlaceService } from "./naverSearch/search.service";
+import { PlaceService } from "./kakaoMap/search.service";
 import { PlaceController } from "./place.controller";
 
 @Module({
